Updated README.md
Revised the file and properly removed the comments for personal usage. Also added further proper acknowledgements. Still need to edit the README.md to include proper examples. Will decide if there needs to be a step-by-step guide with examples on how to run the environment or perhaps just redirect to proper guides because its really just a matter of understanding Python and Django.